# specs/001-build-a-website/contracts/api-specification.yaml
openapi: 3.0.3
info:
  title: Domaine Vallot E-Commerce API
  description: API for boutique winery direct-to-consumer e-commerce website
  version: 1.0.0
  contact:
    name: Domaine Vallot
    email: info@domainevallot.com
servers:
  - url: https://domainevallot.com/api
    description: Production server
  - url: http://localhost:3000/api
    description: Development server
paths:
...
  /cart/{itemId}:
    put:
      summary: Update cart item quantity
      description: Update the quantity of a specific cart item
      security:
        - bearerAuth: []
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - quantity
              properties:
                quantity:
                  type: integer
                  minimum: 1
                  maximum: 12
      responses:
        '200':
          description: Cart item updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'
        '404':
          description: Cart item not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      summary: Remove item from cart
      description: Remove a specific item from the customer's cart
      security:
        - bearerAuth: []
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Item removed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'
        '404':
          description: Cart item not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /orders/{orderId}:
    get:
      summary: Get order details
      description: Retrieve detailed information for a specific order
      security:
        - bearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Order details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDetail'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Product:
      type: object
      properties:
        id:
          type: string
          format: uuid
        sku:
          type: string
        name:
          type: string
        vintage:
          type: integer
        varietal:
          type: string
        price_eur:
          type: number
          format: decimal
        stock_quantity:
          type: integer
        description:
          type: string
        tasting_notes:
          type: string
        food_pairing:
          type: string
        images:
          type: array
          items:
            type: object
            properties:
              url:
                type: string
                format: uri
              alt_text:
                type: string
              is_primary:
                type: boolean
        certifications:
          type: array
          items:
            type: string
            enum: [organic, biodynamic, vegan]
        slug:
          type: string
...
    Cart:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/CartItem'
        subtotal:
          type: number
          format: decimal
        item_count:
          type: integer
    CartItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        product:
          $ref: '#/components/schemas/Product'
        quantity:
          type: integer
        line_total:
          type: number
          format: decimal
    Order:
      type: object
      properties:
        id:
          type: string
          format: uuid
        order_number:
          type: string
        status:
          type: string
          enum: [pending, confirmed, processing, shipped, delivered, cancelled, refunded]
        total_eur:
          type: number
          format: decimal
        created_at:
          type: string
          format: date-time
        shipping_address:
          $ref: '#/components/schemas/Address'
        tracking_number:
          type: string
    OrderDetail:
      allOf:
        - $ref: '#/components/schemas/Order'
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/OrderItem'
            subtotal_eur:
              type: number
              format: decimal
            vat_amount_eur:
              type: number
              format: decimal
            shipping_cost_eur:
              type: number
              format: decimal
            billing_address:
              $ref: '#/components/schemas/Address'
    OrderItem:
      type: object
      properties:
        product_id:
          type: string
          format: uuid
        product_name:
          type: string
        quantity:
          type: integer
        unit_price_eur:
          type: number
          format: decimal
        line_total_eur:
          type: number
          format: decimal
    Address:
      type: object
      required:
        - first_name
        - last_name
        - address_line1
        - city
        - postal_code
        - country_code
      properties:
        first_name:
          type: string
          maxLength: 50
        last_name:
          type: string
          maxLength: 50
        company:
          type: string
          maxLength: 100
        address_line1:
          type: string
          maxLength: 100
        address_line2:
          type: string
          maxLength: 100
        city:
          type: string
          maxLength: 50
        state_province:
          type: string
          maxLength: 50
        postal_code:
          type: string
          maxLength: 20
        country_code:
          type: string
          pattern: '^[A-Z]{2}$'
        phone:
          type: string
          maxLength: 20
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object